First Rough Draft for integration testing support
