+++ rsky-pds/Cargo.toml
@@ -64,11 +64,13 @@ time = "^0.3.36"
 url = "2.5.2"
 async-event-emitter = "0.1.3"
 event-emitter-rs = "0.1.4"
-webpki-roots = { version = "0.26.0-alpha.1" }
 lexicon_cid = { workspace = true }
-once_cell = "1.19.0"
 tracing = "0.1.41"
 tracing-subscriber = "0.3.19"
+
+[dev-dependencies]
+testcontainers = "0.23.2"
+testcontainers-modules = { version = "0.11.6", features = ["postgres", "blocking"] }
 
 [dependencies.rocket_sync_db_pools]
 version = "=0.1.0"
