+++ rsky-pds/src/main.rs
@@ -1,454 +1,6 @@
-#[macro_use]
-extern crate rocket;
-use atrium_api::client::AtpServiceClient;
-use atrium_xrpc_client::reqwest::ReqwestClientBuilder;
-use diesel::prelude::*;
-use diesel::sql_types::Int4;
-use dotenvy::dotenv;
-use rocket::data::{Limits, ToByteUnit};
-use rocket::fairing::{Fairing, Info, Kind};
-use rocket::figment::{
-    util::map,
-    value::{Map, Value},
-};
-use rocket::http::Header;
-use rocket::http::Status;
-use rocket::response::status;
-use rocket::serde::json::Json;
-use rocket::shield::{NoSniff, Shield};
-use rocket::{Request, Response};
-use rsky_common::env::env_list;
-use rsky_identity::types::{DidCache, IdentityResolverOpts};
-use rsky_identity::IdResolver;
-use rsky_pds::account_manager::AccountManager;
-use rsky_pds::apis::*;
-use rsky_pds::config::env_to_cfg;
-use rsky_pds::crawlers::Crawlers;
-use rsky_pds::db::DbConn;
-use rsky_pds::read_after_write::viewer::{LocalViewer, LocalViewerCreatorParams};
-use rsky_pds::sequencer::Sequencer;
-use rsky_pds::well_known::well_known;
-use rsky_pds::{
-    SharedATPAgent, SharedIdResolver, SharedLocalViewer, SharedSequencer, APP_USER_AGENT,
-};
-use std::env;
-use tokio::sync::RwLock;
+use rsky_pds::build_rocket;
 
-pub struct CORS;
-
-#[get("/")]
-pub async fn index() -> &'static str {
-    r#"
-    .------..------..------..------.
-    |R.--. ||S.--. ||K.--. ||Y.--. |
-    | :(): || :/\: || :/\: || (\/) |
-    | ()() || :\/: || :\/: || :\/: |
-    | '--'R|| '--'S|| '--'K|| '--'Y|
-    `------'`------'`------'`------'
-    .------..------..------.
-    |P.--. ||D.--. ||S.--. |
-    | :/\: || :/\: || :/\: |
-    | (__) || (__) || :\/: |
-    | '--'P|| '--'D|| '--'S|
-    `------'`------'`------'
-    
-    This is an atproto [https://atproto.com] Personal Data Server (PDS) running the rsky-pds codebase [https://github.com/blacksky-algorithms/rsky]
-
-    Most API routes are under /xrpc/
-    "#
-}
-
-#[get("/robots.txt")]
-pub async fn robots() -> &'static str {
-    "# Hello!\n\n# Crawling the public API is allowed\nUser-agent: *\nAllow: /"
-}
-
-#[tracing::instrument(skip_all)]
-#[get("/xrpc/_health")]
-pub async fn health(
-    connection: DbConn,
-) -> Result<
-    Json<rsky_pds::models::ServerVersion>,
-    status::Custom<Json<rsky_pds::models::ErrorMessageResponse>>,
-> {
-    let result = connection
-        .run(move |conn| {
-            diesel::select(diesel::dsl::sql::<Int4>("1")) // SELECT 1;
-                .load::<i32>(conn)
-                .map(|v| v.into_iter().next().expect("no results"))
-        })
-        .await;
-    match result {
-        Ok(_) => {
-            let env_version = env::var("VERSION").unwrap_or("0.3.0-beta.3".into());
-            let version = rsky_pds::models::ServerVersion {
-                version: env_version,
-            };
-            Ok(Json(version))
-        }
-        Err(error) => {
-            tracing::error!("Internal Error: {error}");
-            let internal_error = rsky_pds::models::ErrorMessageResponse {
-                code: Some(rsky_pds::models::ErrorCode::ServiceUnavailable),
-                message: Some(error.to_string()),
-            };
-            Err(status::Custom(
-                Status::ServiceUnavailable,
-                Json(internal_error),
-            ))
-        }
-    }
-}
-
-#[catch(default)]
-async fn default_catcher() -> Json<rsky_pds::models::ErrorMessageResponse> {
-    let internal_error = rsky_pds::models::ErrorMessageResponse {
-        code: Some(rsky_pds::models::ErrorCode::InternalServerError),
-        message: Some("Internal error.".to_string()),
-    };
-    Json(internal_error)
-}
-
-/// Catches all OPTION requests in order to get the CORS related Fairing triggered.
-#[options("/<_..>")]
-async fn all_options() {
-    /* Intentionally left empty */
-}
-
-#[rocket::async_trait]
-impl Fairing for CORS {
-    fn info(&self) -> Info {
-        Info {
-            name: "Add CORS headers to responses",
-            kind: Kind::Response,
-        }
-    }
-
-    async fn on_response<'r>(&self, _request: &'r Request<'_>, response: &mut Response<'r>) {
-        response.set_header(Header::new("Access-Control-Allow-Origin", "*"));
-        response.set_header(Header::new(
-            "Access-Control-Allow-Methods",
-            "POST, GET, PATCH, OPTIONS, DELETE",
-        ));
-        response.set_header(Header::new("Access-Control-Allow-Headers", "*"));
-        response.set_header(Header::new("Access-Control-Allow-Credentials", "true"));
-    }
-}
-
-#[launch]
-async fn rocket() -> _ {
-    dotenv().ok();
-
-    let subscriber = tracing_subscriber::FmtSubscriber::new();
-    tracing::subscriber::set_global_default(subscriber).unwrap();
-
-    let db_url = env::var("DATABASE_URL").unwrap_or("".into());
-
-    let db: Map<_, Value> = map! {
-        "url" => db_url.into(),
-        "pool_size" => 20.into(),
-        "timeout" => 30.into(),
-    };
-
-    let figment = rocket::Config::figment()
-        .merge(("databases", map!["pg_db" => db]))
-        .merge(("limits", Limits::default().limit("file", 100.mebibytes())));
-    let cfg = env_to_cfg();
-
-    let sequencer = SharedSequencer {
-        sequencer: RwLock::new(Sequencer::new(
-            Crawlers::new(cfg.service.hostname.clone(), cfg.crawlers.clone()),
-            None,
-        )),
-    };
-    let mut background_sequencer = sequencer.sequencer.write().await.clone();
-    tokio::spawn(async move { background_sequencer.start().await });
-
-    let aws_sdk_config = aws_config::from_env()
-        .endpoint_url(env::var("AWS_ENDPOINT").unwrap_or("localhost".to_owned()))
-        .load()
-        .await;
-
-    let id_resolver = SharedIdResolver {
-        id_resolver: RwLock::new(IdResolver::new(IdentityResolverOpts {
-            timeout: None,
-            plc_url: Some(
-                env::var("PDS_DID_PLC_URL").unwrap_or("https://plc.directory".to_owned()),
-            ),
-            did_cache: Some(DidCache::new(None, None)),
-            backup_nameservers: Some(env_list("PDS_HANDLE_BACKUP_NAMESERVERS")),
-        })),
-    };
-
-    // Keeping unused for other config purposes for now.
-    let app_view_agent = match cfg.bsky_app_view {
-        None => SharedATPAgent {
-            app_view_agent: None,
-        },
-        Some(ref bsky_app_view) => {
-            let client = ReqwestClientBuilder::new(bsky_app_view.url.clone())
-                .client(
-                    reqwest::ClientBuilder::new()
-                        .user_agent(APP_USER_AGENT)
-                        .timeout(std::time::Duration::from_millis(1000))
-                        .build()
-                        .unwrap(),
-                )
-                .build();
-            SharedATPAgent {
-                app_view_agent: Some(RwLock::new(AtpServiceClient::new(client))),
-            }
-        }
-    };
-    let local_viewer = SharedLocalViewer {
-        local_viewer: RwLock::new(LocalViewer::creator(LocalViewerCreatorParams {
-            account_manager: AccountManager {},
-            pds_hostname: cfg.service.hostname.clone(),
-            appview_agent: match cfg.bsky_app_view {
-                None => None,
-                Some(ref bsky_app_view) => Some(bsky_app_view.url.clone()),
-            },
-            appview_did: match cfg.bsky_app_view {
-                None => None,
-                Some(ref bsky_app_view) => Some(bsky_app_view.did.clone()),
-            },
-            appview_cdn_url_pattern: match cfg.bsky_app_view {
-                None => None,
-                Some(ref bsky_app_view) => bsky_app_view.cdn_url_pattern.clone(),
-            },
-        })),
-    };
-
-    let shield = Shield::default().enable(NoSniff::Enable);
-
-    rocket::custom(figment)
-        .mount(
-            "/",
-            routes![
-                index,
-                robots,
-                health,
-                com::atproto::admin::delete_account::delete_account,
-                com::atproto::admin::disable_account_invites::disable_account_invites,
-                com::atproto::admin::disable_invite_codes::disable_invite_codes,
-                com::atproto::admin::enable_account_invites::enable_account_invites,
-                com::atproto::admin::get_account_info::get_account_info,
-                com::atproto::admin::get_invite_codes::get_invite_codes,
-                com::atproto::admin::get_subject_status::get_subject_status,
-                com::atproto::admin::send_email::send_email,
-                com::atproto::admin::update_account_password::update_account_password,
-                com::atproto::admin::update_account_email::update_account_email,
-                com::atproto::admin::update_account_handle::update_account_handle,
-                com::atproto::admin::update_subject_status::update_subject_status,
-                com::atproto::identity::resolve_handle::resolve_handle,
-                com::atproto::identity::update_handle::update_handle,
-                com::atproto::repo::apply_writes::apply_writes,
-                com::atproto::repo::create_record::create_record,
-                com::atproto::repo::delete_record::delete_record,
-                com::atproto::repo::describe_repo::describe_repo,
-                com::atproto::repo::get_record::get_record,
-                com::atproto::repo::import_repo::import_repo,
-                com::atproto::repo::list_records::list_records,
-                com::atproto::repo::list_missing_blobs::list_missing_blobs,
-                com::atproto::repo::put_record::put_record,
-                com::atproto::repo::upload_blob::upload_blob,
-                com::atproto::server::confirm_email::confirm_email,
-                com::atproto::server::create_account::server_create_account,
-                com::atproto::server::create_app_password::create_app_password,
-                com::atproto::server::create_invite_code::create_invite_code,
-                com::atproto::server::create_invite_codes::create_invite_codes,
-                com::atproto::server::create_session::create_session,
-                com::atproto::server::deactivate_account::deactivate_account,
-                com::atproto::server::delete_account::delete_account,
-                com::atproto::server::delete_session::delete_session,
-                com::atproto::server::describe_server::describe_server,
-                com::atproto::server::check_account_status::check_account_status,
-                com::atproto::server::activate_account::activate_account,
-                com::atproto::server::get_service_auth::get_service_auth,
-                com::atproto::server::get_account_invite_codes::get_account_invite_codes,
-                com::atproto::server::get_session::get_session,
-                com::atproto::server::list_app_passwords::list_app_passwords,
-                com::atproto::server::refresh_session::refresh_session,
-                com::atproto::server::request_account_delete::request_account_delete,
-                com::atproto::server::request_email_confirmation::request_email_confirmation,
-                com::atproto::server::request_email_update::request_email_update,
-                com::atproto::server::request_password_reset::request_password_reset,
-                com::atproto::server::reset_password::reset_password,
-                com::atproto::server::revoke_app_password::revoke_app_password,
-                com::atproto::server::update_email::update_email,
-                com::atproto::server::reserve_signing_key::reserve_signing_key,
-                com::atproto::sync::get_blob::get_blob,
-                com::atproto::sync::get_blocks::get_blocks,
-                com::atproto::sync::get_latest_commit::get_latest_commit,
-                com::atproto::sync::get_record::get_record,
-                com::atproto::sync::get_repo::get_repo,
-                com::atproto::sync::get_repo_status::get_repo_status,
-                com::atproto::sync::list_blobs::list_blobs,
-                com::atproto::sync::list_repos::list_repos,
-                com::atproto::sync::subscribe_repos::subscribe_repos,
-                app::bsky::actor::get_preferences::get_preferences,
-                app::bsky::actor::get_profile::get_profile,
-                app::bsky::actor::get_profiles::get_profiles,
-                app::bsky::actor::put_preferences::put_preferences,
-                app::bsky::feed::get_actor_likes::get_actor_likes,
-                app::bsky::feed::get_author_feed::get_author_feed,
-                app::bsky::feed::get_feed::get_feed,
-                app::bsky::feed::get_post_thread::get_post_thread,
-                app::bsky::feed::get_timeline::get_timeline,
-                app::bsky::notification::register_push::register_push,
-                bsky_api_get_forwarder,
-                bsky_api_post_forwarder,
-                well_known,
-                all_options
-            ],
-        )
-        .register("/", catchers![default_catcher])
-        .attach(CORS)
-        .attach(DbConn::fairing())
-        .attach(shield)
-        .manage(sequencer)
-        .manage(aws_sdk_config)
-        .manage(id_resolver)
-        .manage(cfg)
-        .manage(local_viewer)
-        .manage(app_view_agent)
-}
-
-#[cfg(test)]
-mod test {
-    use crate::rocket;
-    use rocket::http::{ContentType, Header, Status};
-    use rocket::local::asynchronous::Client;
-    use rsky_lexicon::com::atproto::server::{
-        CreateAccountInput, CreateSessionInput, CreateSessionOutput,
-    };
-
-    // We use a lock to synchronize between tests so FS operations don't race.
-    static FS_LOCK: parking_lot::Mutex<()> = parking_lot::const_mutex(());
-
-    // We use a lock to synchronize between tests so FS operations don't race.
-    static CLIENT_LOCK: parking_lot::Mutex<Option<Client>> = parking_lot::const_mutex(None);
-    static mut CLIENT: Option<Client> = None;
-
-    async unsafe fn client() -> Client {
-        match CLIENT {
-            None => {
-                let x = Client::untracked(rocket().await)
-                    .await
-                    .expect("Rocket Client");
-                CLIENT =  Some(x);
-                x
-            }
-        }
-    }
-
-    fn setup_database() {}
-
-    #[tokio::test]
-    async fn test_create_invite_code() {
-        let client = unsafe { client() }.await;
-        let input = CreateAccountInput {
-            email: Some("testing_user@rsky.com".to_string()),
-            handle: "testing_user.rsky.com".to_string(),
-            did: None,
-            invite_code: None,
-            verification_code: None,
-            verification_phone: None,
-            password: Some("test_password".to_string()),
-            recovery_key: None,
-            plc_op: None,
-        };
-        let response = client
-            .post("/xrpc/com.atproto.server.createAccount")
-            .header(ContentType::JSON)
-            .body(serde_json::to_string(&input).unwrap().into_bytes())
-            .dispatch()
-            .await;
-        let x = response.body();
-        assert_eq!(response.status(), Status::Ok);
-    }
-
-    #[tokio::test]
-    async fn test_create_account() {
-        let client = unsafe { client() }.await;
-
-        let input = CreateAccountInput {
-            email: Some("testing_user@rsky.com".to_string()),
-            handle: "testing_user.rsky.com".to_string(),
-            did: None,
-            invite_code: None,
-            verification_code: None,
-            verification_phone: None,
-            password: Some("test_password".to_string()),
-            recovery_key: None,
-            plc_op: None,
-        };
-        let response = client
-            .post("/xrpc/com.atproto.server.createAccount")
-            .header(ContentType::JSON)
-            .body(serde_json::to_string(&input).unwrap().into_bytes())
-            .dispatch()
-            .await;
-        let x = response.body();
-        assert_eq!(response.status(), Status::Ok);
-    }
-
-    #[tokio::test]
-    async fn test_create_session() {
-        let client = unsafe { client() }.await;
-        let input = CreateSessionInput {
-            identifier: "testing_user.rsky.com".to_string(),
-            password: "testing_password".to_string(),
-        };
-        let mut response = client
-            .post("/xrpc/com.atproto.server.createSession")
-            .body(serde_json::to_string(&input).unwrap().into_bytes())
-            .dispatch()
-            .await;
-        let status = response.status();
-
-        assert_eq!(status, Status::Ok);
-        let body = response.into_json::<CreateSessionOutput>().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_get_session() {
-        let client = unsafe { client() }.await;
-        let mut response = client
-            .get("/xrpc/com.atproto.server.getSession")
-            .header(Header::new("", ""))
-            .dispatch()
-            .await;
-        assert_eq!(response.status(), Status::Ok);
-    }
-
-    #[tokio::test]
-    async fn test_resolve_handle() {
-        let client = unsafe { client() }.await;
-        let mut response = client
-            .get("/xrpc/com.atproto.identity.resolveHandle?ripperoni.com")
-            .dispatch()
-            .await;
-        assert_eq!(response.status(), Status::Ok);
-    }
-
-    #[tokio::test]
-    async fn test_index() {
-        let client = unsafe { client() }.await;
-        let mut response = client.get("/").dispatch().await;
-        assert_eq!(response.status(), Status::Ok);
-    }
-
-    #[tokio::test]
-    async fn test_robots_txt() {
-        let client = unsafe { client() }.await;
-        let mut response = client.get("/robots.txt").dispatch().await;
-        let response_status = response.status();
-        let response_body = response.into_string().await.unwrap();
-        assert_eq!(response_status, Status::Ok);
-        assert_eq!(
-            response_body,
-            "# Hello!\n\n# Crawling the public API is allowed\nUser-agent: *\nAllow: /"
-        );
-    }
+#[rocket::main]
+async fn main() {
+    let _ = build_rocket().await.launch().await;
 }
