--- rsky-pds/tests/integration_test.rs
@@ -1,5 +1,0 @@
-// // tests/integration_tests.rs
-// #[test]
-// fn test_returns_one() {
-//     assert_eq!(returner::return_one(), 1);
-// }
